fetching_teams_meetings/Single_tenant/requirements.txt
Flask
msal
requests
python-dotenv
